% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelib_objects.r
\docType{data}
\name{ddf}
\alias{ddf}
\title{Dummy Data Set}
\description{
A data.frame representing 4 global factors and 12 facets.
}
\author{
Nikolaos Giallousis \email{ngiallousis@psyget.gr}
#@references \url{data_blah.com}
}
\keyword{data}
