% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelib_functions.R
\name{decorate_corr}
\alias{decorate_corr}
\title{Add stars to correlation matrix}
\usage{
decorate_corr(values, signi)
}
\arguments{
\item{values}{matrix the r matrix from rcorr}

\item{signi}{matrix the P matrix from rcorr}
}
\value{
a character matrix
}
