% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelib_functions.R
\name{extract_q}
\alias{extract_q}
\title{Extract a named element from all elements of a nested list object (Standard evaluation escape hatch) .}
\usage{
extract_q(llistObj, element)
}
\description{
Standard evaluation escape hatch for \link{extract}.
}
