% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelib_objects.r
\docType{data}
\name{facet_items}
\alias{facet_items}
\title{Test Ontology}
\description{
A list relating items to facets.
}
\author{
Nikolaos Giallousis \email{ngiallousis@psyget.gr}
#@references \url{data_blah.com}
}
\keyword{data}
