% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelib_objects.r
\docType{data}
\name{factor_facets}
\alias{factor_facets}
\title{Test Ontology}
\description{
A list relating facets to global factors.
}
\author{
Nikolaos Giallousis \email{ngiallousis@psyget.gr}
#@references \url{data_blah.com}
}
\keyword{data}
