% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelib_functions.R
\name{fit_bulk_models}
\alias{fit_bulk_models}
\title{Iteratively Fit and Modify CFA models}
\usage{
fit_bulk_models(dataObj, n = 12, start_model)
}
\examples{

bulk_fit0 = fit_bulk_models(d, 20, model0)
Reduce(rbind, bulk_fit0$fit_stats)
}
