% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelib_functions.R
\name{fitmeasure_select}
\alias{fitmeasure_select}
\title{Select Fit Measures from lavaan output}
\usage{
fitmeasure_select(fitObj, selection = c("chisq", "df", "pvalue", "gfi",
  "agfi", "rmsea", "tli", "cfi", "srmr"))
}
\arguments{
\item{fitObj}{an object fitted with lavaan}

\item{selection}{character the fit indices to select}
}
\value{
a numeric matrix
}
