% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelib_functions.R
\name{form_prop_table}
\alias{form_prop_table}
\title{Format a Proportion's Table}
\usage{
form_prop_table(cat1, cat2)
}
\arguments{
\item{cat1}{factor A categorical variable}

\item{cat2}{factor A categorical variable}

\item{output}{character Desired output, either \code{pander}, \code{xtable}, or \code{default}}
}
\value{
A formated table
}
