% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelib_functions.R
\name{func2matx}
\alias{func2matx}
\title{Call a function on a nested list objects and bind the result into a matrix, row- or column-wise.}
\usage{
func2matx(bind, llistObj, func, ...)
}
\arguments{
\item{bind}{The binding function, i.e. \code{\link{cbind}} or \code{\link{rbind}}.}

\item{llistObj}{A nested list object with two levels.}

\item{func}{The function to \emph{call} (not \link{apply}) on llistObj.}

\item{...}{Other arguments to be passed to \code{func}.}
}
\value{
A matrix of the return value of func called on llistObj.
}
\details{
\code{func2matx} calls rather than applies the function on the nested list. If you want to use the apply family you can use, for instance, \code{\link{lapply}} a closure right away in the \code{func} argument.
}
\examples{
nested_list = list(list(a=c('a','b','c'), b=1:3), list(a=c('d','e','f'),b=4:6))
func2matx(rbind, nested_list, extract, element=a)

func2matx(rbind, nested_list, paste, sep='-', collaspe='\\n')
}
\author{
Nikolaos Giallousis \email{ngiallousis@psyget.gr}
}
