% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelib_functions.R
\name{llistify}
\alias{llistify}
\title{Turn a flat list to a llistObj}
\usage{
llistify(global, flat)
}
\arguments{
\item{global}{list a factor_facet list}

\item{flat}{list a list, most plausibly a list of analysis by facet}
}
\details{
Useful when there are global dimensions, i.e. hierarchy in scales
}
