% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelib_functions.R
\name{modif}
\alias{modif}
\title{Return most important modification indices from lavaan object}
\usage{
modif(fitObj, many = 10)
}
\arguments{
\item{fitObj}{an object fitted with lavaan}

\item{many}{integer how many indices to use}
}
