% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelib_functions.R
\name{myExclude}
\alias{myExclude}
\title{Global Item Exclusion}
\usage{
myExclude(item_list, exclude_list, num = F, col_str = "Q")
}
\details{
Nested list objects, as usual, structured by Subscale
Uses \code{\link{Map}} to exclude
TODO deal with numeric vs. colnames item definition
.. if this is an option, it should read the colnames properly
TODO deal with flat list vs nested list (facets vs global>facets)
TODO check lengths so that not too many items are deleted
QUEST is there order in the POC lists?
}
