% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelib_functions.R
\name{pool_mlm}
\alias{pool_mlm}
\title{(Tweak) Combine Multiple Imputation Results}
\usage{
pool_mlm(results, variances, call = sys.call(), df.complete = Inf, ...)
}
\description{
cy_diag <- reg_diagnosis(cy_lm)
ex_diag <- reg_diagnosis(ex_lm)
pe_diag <- reg_diagnosis(pe_lm)
}
\details{
par(mfrow=c(2,4))
reg_diagnosis_summary(cy_diag)
reg_diagnosis_summary(ex_diag)
reg_diagnosis_summary(pe_diag)

summary(cy_diag$original)
summary(cy_diag$updated)
summary(ex_diag$original)
summary(ex_diag$updated)
summary(pe_diag$original)
summary(pe_diag$updated)

cy_lm2 <- update(cy_diag$updated, . ~ . -practice -location -nyear -nempl,
                 data=cy_diag$xdata)
ex_lm2 <- update(ex_diag$updated, . ~ . -age -location -nempl,
                 data=ex_diag$xdata)
pe_lm2 <- update(pe_diag$updated, . ~ . -age -practice -location,
                 data=pe_diag$xdata)

cy_fit16 <- fitted(cy_lm2)
ex_fit16 <- fitted(ex_lm2)
pe_fit16 <- fitted(pe_lm2)
}
